+++ tools/legacy/sample_misc/wayland/src/class_wayland.cpp
@@ -61,6 +61,7 @@ Wayland::Wayland()
           m_x(0),
           m_y(0),
           m_perf_mode(false),
+          m_requiredTiled4(false),
           m_buffers_list() {
     std::memset(&m_poll, 0, sizeof(m_poll));
 }
@@ -364,6 +365,11 @@ struct wl_buffer* Wayland::CreatePrimeBuffer(uint32_t name,
         int i                                            = 0;
         uint64_t modifier                                = I915_FORMAT_MOD_Y_TILED;
 
+    #if defined(DRM_LINUX_MODIFIER_TILED4_SUPPORT)
+        if (m_requiredTiled4)
+            modifier = I915_FORMAT_MOD_4_TILED;
+    #endif
+
         dmabuf_params = zwp_linux_dmabuf_v1_create_params(m_dmabuf);
         for (i = 0; i < 2; i++) {
             zwp_linux_buffer_params_v1_add(dmabuf_params,
@@ -459,9 +465,13 @@ void Wayland::RegistryGlobal(struct wl_registry* registry,
         wl_drm_add_listener(m_drm, &drm_listener, this);
     }
 #if defined(WAYLAND_LINUX_DMABUF_SUPPORT)
-    else if (msdk_match(interface, "zwp_linux_dmabuf_v1"))
+    else if (msdk_match(interface, "zwp_linux_dmabuf_v1")) {
+        static const struct zwp_linux_dmabuf_v1_listener dmabuf_listener = { dmabuf_format,
+                                                                             dmabuf_modifier };
         m_dmabuf = static_cast<zwp_linux_dmabuf_v1*>(
-            wl_registry_bind(registry, name, &zwp_linux_dmabuf_v1_interface, version));
+            wl_registry_bind(registry, name, &zwp_linux_dmabuf_v1_interface, 3));
+        zwp_linux_dmabuf_v1_add_listener(m_dmabuf, &dmabuf_listener, this);
+    }
 #endif
 }